requirements.txt
pywhatkit>=5.5
Pillow>=9.0.0
numpy>=1.21.0
